+++ pdoc/__init__.py
@@ -390,7 +390,7 @@ You can find an example in [`examples/library-usage`](https://github.com/mitmpro
 from __future__ import annotations
 
 __docformat__ = "markdown"  # explicitly disable rST processing in the examples above.
-__version__ = "10.0.4"  # this is read from setup.py
+__version__ = "11.0.0"  # this is read from setup.py
 
 import traceback
 import warnings